Update implementation, based on the following changes:
Changes in Revision 1

 * Moved the concept of Variable Substitution to be a responsibility of the
package manager, meaning any variables must be substituted before the files are
presented to the build system.

 * Introduced a section related to alternative parsing options for the same
interface file. Additionally, the bmi filename now includes a fragment to
identify which of the alternative parsings are being used.

 * Clarified the “Search order” section to explain it in better detail
including the reason for it.

 * Made it a requirement to provide a meta-ixx-info file.

 * Proof-of-concept implementation is now required to generate recipes for
building each individual bmi file, since names are now dependent on the
contents of the meta-ixx-info file.